Read some configuration from environment
Environment variables are read from a `config.env` file, both when
running the dev server directly, as well as using docker-compose. The
django-environ package is used as a helper in this.

In the Dockerfile changed the container to be started with an entrypoint
script. This is needed so that the database migration command gets run
with the correct environment variables too.

HP-213