HP-644 Add GDPR API endpoint
Helsinki GDPR API Library is added to provide the basic functionality which we customize a bit to fit this example.

UserData model is chosen to be the GDPR API Model and modified to serialize pet_name.

If the profile backend should query the endpoint it would receive the following JSON:

  {
    "key": "USERDATA",
    "children": [
      {
        "key": "PET_NAME",
        "value": "An example name"
      }
    ]
  }